Former-commit-id: ba6b2de4807dd80b08a63cf61907023954eea676
Former-commit-id: dffed09b1486af8ffd105e22c29c4d5e9907a846
Former-commit-id: f32e945c059766c57c939fbb7de94c5e61b15d0a

+++ src/ImageProcessor/Samplers/Resamplers/CatmullRomResampler.cs
@@ -12,13 +12,13 @@ namespace ImageProcessor.Samplers
     public class CatmullRomResampler : IResampler
     {
         /// <inheritdoc/>
-        public double Radius => 2;
+        public float Radius => 2;
 
         /// <inheritdoc/>
-        public double GetValue(double x)
+        public float GetValue(float x)
         {
-            const double B = 0;
-            const double C = 1 / 2d;
+            const float B = 0;
+            const float C = 1 / 2f;
 
             return ImageMaths.GetBcValue(x, B, C);
         }
